Integrated contig ends merger. Integrated evidence proportion to edges cost

+++ src/ngsep/assembly/NStatisticsCalculator.java
@@ -6,11 +6,11 @@ import java.util.Collections;
 import java.util.List;
 
 public class NStatisticsCalculator {
-	public static int[] calculateNStatistics(List<Integer> numbers) {
-		int[] answer = new int [10];
+	public static long[] calculateNStatistics(List<Integer> numbers) {
+		long [] answer = new long [10];
 		Arrays.fill(answer, 0);
 		Collections.sort(numbers,(l1,l2)-> l2-l1);
-		int total = 0;
+		long total = 0;
 		for(int i:numbers) total+=i;
 		answer[0] = 0;
 		double current = 0;
@@ -24,7 +24,7 @@ public class NStatisticsCalculator {
 		}
 		return answer;
 	}
-	public static void printNStatistics(int [] stats, PrintStream out) {
+	public static void printNStatistics(long [] stats, PrintStream out) {
 		for(int i=1;i<stats.length;i++) {
 			out.println("N"+(10*i)+"\t"+stats[i]);
 		}
